Added base operation fee charge EET clients.

+++ hapi-fees/src/main/java/com/hedera/services/usage/token/meta/ExtantTokenContext.java
@@ -54,7 +54,7 @@ public class ExtantTokenContext {
 
 
 	public long getExistingRbSize() {
-		return existingNameLen + existingSymLen + existingMemoLen + existingAdminKeyLen
+		return (long)existingNameLen + existingSymLen + existingMemoLen + existingAdminKeyLen
 				+ existingKycKeyLen + existingFreezeKeyLen + existingWipeKeyLen
 				+ existingFreezeKeyLen;
 	}
